DCP Backfill: Use size_t instead of uint32_t to record mem usage
We keep a record of DCP backfill memory usage to ensure that backfill
memory usage is bounded by a finite size (backfill buffer size).

We must consistently use size_t instead of uint32_t so that there is
no overflow error.

Reviewed-on: http://review.couchbase.org/77270
Tested-by: Build Bot <build@couchbase.com>

+++ src/dcp/backfill-manager.cc
@@ -162,7 +162,7 @@ void BackfillManager::schedule(VBucket& vb,
     ExecutorPool::get()->schedule(managerTask);
 }
 
-bool BackfillManager::bytesRead(uint32_t bytes) {
+bool BackfillManager::bytesRead(size_t bytes) {
     LockHolder lh(lock);
     if (scanBuffer.itemsRead >= scanBuffer.maxItems) {
         return false;
@@ -192,7 +192,7 @@ bool BackfillManager::bytesRead(uint32_t bytes) {
     return true;
 }
 
-void BackfillManager::bytesSent(uint32_t bytes) {
+void BackfillManager::bytesSent(size_t bytes) {
     LockHolder lh(lock);
     if (bytes > buffer.bytesRead) {
         throw std::invalid_argument("BackfillManager::bytesSent: bytes "
