+++ src/dcp/producer.h
@@ -138,8 +138,8 @@ enum class MutationType {
     void notifyStreamReady(uint16_t vbucket);
 
     void notifyBackfillManager();
-    bool recordBackfillManagerBytesRead(uint32_t bytes);
-    void recordBackfillManagerBytesSent(uint32_t bytes);
+    bool recordBackfillManagerBytesRead(size_t bytes);
+    void recordBackfillManagerBytesSent(size_t bytes);
     void scheduleBackfillManager(VBucket& vb,
                                  const active_stream_t& s,
                                  uint64_t start,
